+++ src/BigRational.php
@@ -6,7 +6,6 @@ namespace Brick\Math;
 
 use Brick\Math\Exception\DivisionByZeroException;
 use Brick\Math\Exception\MathException;
-use Brick\Math\Exception\NumberFormatException;
 use Brick\Math\Exception\RoundingNecessaryException;
 use InvalidArgumentException;
 use LogicException;
@@ -67,33 +66,6 @@ final readonly class BigRational extends BigNumber
 
         $this->numerator = $numerator;
         $this->denominator = $denominator;
-    }
-
-    /**
-     * Creates a BigRational out of a numerator and a denominator.
-     *
-     * If the denominator is negative, the signs of both the numerator and the denominator
-     * will be inverted to ensure that the denominator is always positive.
-     *
-     * @deprecated Use ofFraction() instead.
-     *
-     * @param BigNumber|int|float|string $numerator   The numerator. Must be convertible to a BigInteger.
-     * @param BigNumber|int|float|string $denominator The denominator. Must be convertible to a BigInteger.
-     *
-     * @throws NumberFormatException      If an argument does not represent a valid number.
-     * @throws RoundingNecessaryException If an argument represents a non-integer number.
-     * @throws DivisionByZeroException    If the denominator is zero.
-     */
-    public static function nd(
-        BigNumber|int|float|string $numerator,
-        BigNumber|int|float|string $denominator,
-    ): BigRational {
-        trigger_error(
-            'The BigRational::nd() method is deprecated, use BigRational::ofFraction() instead.',
-            E_USER_DEPRECATED,
-        );
-
-        return self::ofFraction($numerator, $denominator);
     }
 
     /**