+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## UNRELEASED (0.15.0)
+
+💥 **Breaking changes**
+
+- deprecated method `BigRational::nd()` has been removed, use `ofFraction()` instead
+
 ## [0.14.8](https://github.com/brick/math/releases/tag/0.14.8) - 2026-02-10
 
 🗑️ **Deprecations**